+++ shoekream/src/main/java/com/supreme/shoekream/controller/page/ProductPageController.java
@@ -1,0 +1,18 @@
+package com.supreme.shoekream.controller.page;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.servlet.ModelAndView;
+
+@Controller
+@RequestMapping("product") //http://localhost:8888/product
+public class ProductPageController {
+
+    @GetMapping(path="") //http://localhost:8888/product
+    public ModelAndView product(){
+        return new ModelAndView("/product");
+    }   //viewName: 페이지이름이랑 같아야함
+
+
+}
