+++ shoekream/src/main/resources/templates/product.html
@@ -167,7 +167,7 @@
                                             <div class="detail_main_title lg">
                                                 <div class="main_title_box">
                                                     <div>
-                                                    <a href="/brands/nike" click.prevent="onClickBrand" class="brand" >Nike</a>
+                                                    <a href="/brand" click.prevent="onClickBrand" class="brand" >Nike</a>
                                                     </div>
                                                     <p class="title">Nike Dunk Low SE Seoul</p>
                                                     <p class="sub_title">나이키 덩크 로우 SE 서울</p>
@@ -1499,9 +1499,6 @@
                                                         <!---->
                                                     </div>
                                                 </div>
-                                                <div class="img_count">
-                                                    <span class="count">+5</span>
-                                                </div>
                                             </a>
                                         </div>
                                         <div class="feed_card item square" style="position: absolute; left: 614px; top: 0px;">
@@ -1674,7 +1671,7 @@
                                         </div>
                                     </div>
                                     <div class="more_btn_box">
-                                        <button type="button" class="btn outlinegrey medium" onclick="location.href='./social/social_product.html'">더보기</button>
+                                        <button type="button" class="btn outlinegrey medium" onclick="location.href='/social/social_product'">더보기</button>
                                     </div>
                                     <div class="list_loading" style="display: none;">
                                         <img src="/_nuxt/img/loading.410eb77.gif" loading="lazy" alt="리스트 로딩중입니다." class="loading_img">
@@ -1688,7 +1685,7 @@
                                 <h3 class="more_title">
                                     <span class="owner">Nike</span>
                                     <span class="title_txt">의 다른 상품</span>
-                                    <a href="./brands.html" class="btn_more">
+                                    <a href="/brand" class="btn_more">
                                         <span class="btn_txt">더보기</span>
                                         <img src="/img/right_arrow.png" class="icon sprite-icons arr-right-gray"></img>
                                     </a>
